rename `table_collection.rs` to `master_table.rs`
