+++ triton-vm/benches/prove_fib_100.rs
@@ -1,12 +1,16 @@
-use criterion::{criterion_group, criterion_main, BenchmarkId, Criterion};
-
+use criterion::criterion_group;
+use criterion::criterion_main;
+use criterion::BenchmarkId;
+use criterion::Criterion;
 use triton_profiler::prof_start;
 use triton_profiler::prof_stop;
-use triton_profiler::triton_profiler::{Report, TritonProfiler};
+use triton_profiler::triton_profiler::Report;
+use triton_profiler::triton_profiler::TritonProfiler;
+
 use triton_vm::instruction::sample_programs;
 use triton_vm::proof::Claim;
 use triton_vm::stark::Stark;
-use triton_vm::table::table_collection::MasterBaseTable;
+use triton_vm::table::master_table::MasterBaseTable;
 use triton_vm::vm::Program;
 
 /// cargo criterion --bench prove_fib_100
