+++ triton-vm/src/op_stack.rs
@@ -1,9 +1,12 @@
-use super::error::{vm_fail, InstructionError::*};
-use super::ord_n::{Ord16, Ord16::*};
 use anyhow::Result;
 use num_traits::Zero;
 use twenty_first::shared_math::b_field_element::BFieldElement;
 use twenty_first::shared_math::x_field_element::XFieldElement;
+
+use super::error::vm_fail;
+use super::error::InstructionError::*;
+use super::ord_n::Ord16;
+use super::ord_n::Ord16::*;
 
 #[derive(Debug, Clone)]
 pub struct OpStack {
@@ -125,7 +128,6 @@ impl OpStack {
 
 #[cfg(test)]
 mod op_stack_test {
-
     use twenty_first::shared_math::b_field_element::BFieldElement;
 
     use crate::{op_stack::OpStack, ord_n::Ord16};
