+++ triton-vm/src/table.rs
@@ -6,9 +6,9 @@ pub mod extension_table;
 pub mod hash_table;
 pub mod instruction_table;
 pub mod jump_stack_table;
+pub mod master_table;
 pub mod op_stack_table;
 pub mod processor_table;
 pub mod program_table;
 pub mod ram_table;
-pub mod table_collection;
 pub mod table_column;
